Introduce BUILD_CACHE_NAME variable
Packages for 20180129 are hardcoded for /opt/spack_20180129 path which
leads to error in some cases.

Hence we need to start with a clean slate for the visionary branch.

BUILD_CACHE_NAME controls which build_cache folder to use.

By default BUILD_CACHE_NAME=visionary, so the build_cache resides in
~vis_jenkins/build_caches/visionary.

In the future nightly jobs can be run with a different BUILD_CACHE_NAME
in order to allow for a clean rebuild.

Change-Id: I17761e2e834d7126cee8b3178bc2471f91ea3e64