Generate /opt/init/modules.sh dynamically
Since typically we have one version of environment-modules installed per
container that does not change, but the location of the init-scripts
folder within environment-modules is subject to change (it did so from
3.x to 4.x), we generate the /opt/init/modules.sh dynmaically after
build.

We hardcode the folder in which the init scripts reside into the script.
If the path should change for any reason (e.g. bind mounting a different
spack version or whatever), it could be easily regenerated via
/opt/spack_install_scripts/generate_modules.sh

Change-Id: I9bae12be665adf84edca65af91440c4115df170b