Add short app names by dropping visionary- prefix
This will duplicate apps for the time being, but since they just consist
of a view scripts it is not much overhead.

This will ease usage on the terminal because the superfuous visionary-
prefix does not have to be typed all the time.

Still, for better consistency between spack packages and views, we keep
the "visionary-"-prefix for folders under /opt/spack_views.

Change-Id: Ic5e19ce0d2379096122119437c52bb64baac93dd