Fix 'illegal instruction' error on AMTHosts
We have several different architectures in the cluster:
* AMTHosts run on Sandybridge
* HBPHosts run on Haswell

The latter support more instruction sets than the former, most notably
AVX2.

The newest upstream merge also merges this commit
3c4322bf1abb7af691179434652188b64e90e4dc which introduced custom targets
per spec via `target=` as well as automatic architecture detection.

The container build executor runs on Haswell with AVX2 supported, hence
all packages were built with it. If almost any program in the container
is then executed on AMTHosts, we get an `illegal instruction` error
because AVX2 is not supported.

As a quick hotfix we have to pin the target for all installed packages
to `sandybridge`. In the future the aforementioned commit will allow us
to build several variants of each package and activate the correct one
based on host-architecture when the container is run.

Depends-On: 8388

Change-Id: Ic0211c22936fcaec0465a8b33d279f23a231ca71