Allow spack changeset via gerrit comment trigger
Since often times yashchiki and spack changes are tested together but
have no real dependency on one another, we misuse the `Depends-On`
mechanism in the commit message to build a container with a specific
spack and yashchiki changeset.

This changeset adds the possibility to specify:
* `WITH_SPACK_CHANGE=<change-num>` to use the latest patch set of the
  given spack changeset for the build
* `WITH_SPACK_REFSPEC=<refspec>` to specify a complete spack refspec
  that is to be used for this build (i.e.,
  refs/changes/<change-num[-2:]>/<change-num>/<patch-level>) to have
  full control over which changeset/patch level to build.

These take priority over commit-specified `Depends-On:` and are mutually
exclusive with jenkins-specified build parameters since each build gets
either triggered manually in jenkins or via gerrit.

Change-Id: Id3b41ae9df4b636256edd54e9b173bef9632dae2