Include path to cache in case of build error
When a container build fails, all packages that were successfully built
up to that point are stored in a fail-cache.

If the same gerrit change gets built again, that cache is used
automatically to reduce turn-around.

Sometimes, work shifts from one gerrit change to another, but the user
still wishes to preserve his build-progress.

Hence, this change makes it explicit which cache was created in case of
build errors.

Change-Id: I6842ba256cc94e5e3c92ff772c8e3de10d380822