Fix array expansion in HEREDOC
If an array is empty and its expanded in a HEREDOC, regular quotation
rules do not apply.

This results in an empty array being expanded as "" which in turn causes
spack's argument parser to panic because it does not recognize "" as a
valid command -> the build fails :(

Fix it! Fix it! Fix it!

Change-Id: I52641809ea61815251010a6077847bdf20b0c21e